Move finding package dependencies to own module
This is in preparation of moving packageDependencies to configuration,
which will allow us to implement special behavior for the meteor
environment.
