+++ lib/Configuration.js
@@ -7,6 +7,7 @@ import semver from 'semver';
 
 import FileUtils from './FileUtils';
 import JsModule from './JsModule';
+import findPackageDependencies from './findPackageDependencies';
 import meteorEnvironment from './environments/meteorEnvironment';
 import nodeEnvironment from './environments/nodeEnvironment';
 import normalizePath from './normalizePath';
@@ -327,22 +328,8 @@ export default class Configuration {
   }
 
   packageDependencies(): Array<string> {
-    const packageJson = FileUtils.readJsonFile(
-      path.join(this.workingDirectory, 'package.json'));
-    if (!packageJson) {
-      return [];
-    }
-
-    const keys = ['dependencies', 'peerDependencies'];
-    if (this.get('importDevDependencies')) {
-      keys.push('devDependencies');
-    }
-    const result = [];
-    keys.forEach((key: string) => {
-      if (Object.prototype.hasOwnProperty.call(packageJson, key)) {
-        result.push(...Object.keys(packageJson[key]));
-      }
-    });
-    return result;
+    return findPackageDependencies(
+      this.workingDirectory,
+      this.get('importDevDependencies'));
   }
 }
